moved import replay operator to seperate file
