+++ addon/import_vcap/operators.py
@@ -10,19 +10,6 @@ from bpy.props import StringProperty, BoolProperty, EnumProperty
 from bpy_extras.io_utils import ImportHelper, ExportHelper
 import bpy
 
-
-def read_some_data(context, filepath, use_some_setting):
-    print("running read_some_data...")
-    f = open(filepath, 'r', encoding='utf-8')
-    data = f.read()
-    f.close()
-
-    # would normally load the data here
-    print(data)
-
-    return {'FINISHED'}
-
-
 # ImportHelper is a helper class, defines filename and
 # invoke() function which calls the file selector.
 
@@ -116,70 +103,6 @@ class ImportEntityOperator(Operator, ImportHelper):
             entity.load_entity(file, context, context.scene.collection)
         return {'FINISHED'}
 
-class ImportReplayOperator(Operator, ImportHelper):
-    bl_idname = "vcap.importreplay"
-    bl_label = "Import Minecraft Replay"
-
-    # ImportHelper mixin class uses this
-    filename_ext = ".txt"
-
-    filter_glob: StringProperty(
-        default="*.replay",
-        options={'HIDDEN'},
-        maxlen=255,  # Max internal buffer length, longer would be clamped.
-    )
-    
-    import_world: BoolProperty(
-        name="Import World",
-        description="Import world blocks (significantly increases import time.)",
-        default=True
-    )
-        
-    import_entities: BoolProperty(
-        name="Import Entities",
-        description="Import Minecraft entities and their animations.",
-        default=True
-    )
-
-    separate_parts: BoolProperty(
-        name="Separate Entity Parts",
-        description="Import every ModelPart in an entity as a seperate object. Only works on multipart entities.",
-        default=False
-    )
-    
-    use_vertex_colors: BoolProperty(
-        name="Use Block Colors",
-        description="Import block colors from the file (grass tint, etc). If unchecked, world may look very grey.",
-        default=True,
-    )
-    
-    merge_verts: BoolProperty(
-        name="Merge Vertices",
-        description="Run a 'merge by distance' operation on the imported world. May exhibit unpredictable behavior.",
-        default=False
-    )
-
-    hide_entities: BoolProperty(
-        name="Auto-hide Entities",
-        description="Hide entities before they've been spawned and after they've been killed.",
-        default=True
-    )
-
-    def execute(self, context: Context):
-        settings = replay_file.ReplaySettings(
-            world=self.import_world,
-            entities=self.import_entities,
-            separate_parts=self.separate_parts,
-            hide_entities=self.hide_entities,
-
-            vcap_settings=VCAPSettings(
-                use_vertex_colors=self.use_vertex_colors,
-                merge_verts=self.merge_verts
-            )
-        )
-        replay_file.load_replay(self.filepath, context, context.scene.collection, self, settings=settings)
-        return {'FINISHED'}
-
 class ExportCameraXMLOperator(Operator, ExportHelper):
     bl_idname = "vcap.exportcameraxml"
     bl_label = "Export Camera XML"
@@ -212,10 +135,6 @@ def menu_func_import2(self, context):
     self.layout.operator(ImportEntityOperator.bl_idname,
                          text="Test Replay Entity (.xml)")
 
-def menu_func_replay(self, context):
-    self.layout.operator(ImportReplayOperator.bl_idname,
-                         text="Minecraft Replay File (.replay)")
-
 def menu_func_camera_xml(self, context):
     self.layout.operator(ExportCameraXMLOperator.bl_idname,
                          text="Camera Animation (.xml)")
@@ -223,22 +142,18 @@ def menu_func_camera_xml(self, context):
 def register():
     bpy.utils.register_class(ImportVcap)
     bpy.utils.register_class(ImportEntityOperator)
-    bpy.utils.register_class(ImportReplayOperator)
     bpy.utils.register_class(ExportCameraXMLOperator)
     bpy.types.TOPBAR_MT_file_import.append(menu_func_import)
     # bpy.utils.register_class(ImportTestOperator)
     bpy.types.TOPBAR_MT_file_import.append(menu_func_import2)
-    bpy.types.TOPBAR_MT_file_import.append(menu_func_replay)
     bpy.types.TOPBAR_MT_file_export.append(menu_func_camera_xml)
 
 
 def unregister():
     bpy.utils.unregister_class(ImportVcap)
     bpy.utils.unregister_class(ImportEntityOperator)
-    bpy.utils.unregister_class(ImportReplayOperator)
     bpy.utils.unregister_class(ExportCameraXMLOperator)
     bpy.types.TOPBAR_MT_file_import.remove(menu_func_import)
     # bpy.utils.unregister_class(ImportTestOperator)
     bpy.types.TOPBAR_MT_file_import.remove(menu_func_import2)
-    bpy.types.TOPBAR_MT_file_import.remove(menu_func_replay)
     bpy.types.TOPBAR_MT_file_export.remove(menu_func_camera_xml)
